refactor: split MainActivity.kt in separate files
Split MainActivity.kt file into separate files each the respective
components instead of having one massive file.
